separate file for array methods
